Updated repo with new codesample.
