--- src/DatabaseMirroringProject.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30320.27
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatabaseMirroringProject", "DatabaseMirroringProject\DatabaseMirroringProject.csproj", "{B801C0A8-1938-48BF-9AC5-ACE60EAB6605}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B801C0A8-1938-48BF-9AC5-ACE60EAB6605}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B801C0A8-1938-48BF-9AC5-ACE60EAB6605}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B801C0A8-1938-48BF-9AC5-ACE60EAB6605}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B801C0A8-1938-48BF-9AC5-ACE60EAB6605}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8924EEC3-D4B6-4514-A86C-D7268C7E3D57}
-	EndGlobalSection
-EndGlobal
