--- src/DatabaseMirroringProject/App_Data/Production/SuperOfficeFederatedLogin.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDqzCCApMCAQgwDQYJKoZIhvcNAQEFBQAwgZMxCzAJBgNVBAYTAk5PMQ0wCwYD
-VQQIDARPc2xvMQ0wCwYDVQQHDARPc2xvMRcwFQYDVQQKDA5TdXBlck9mZmljZSBB
-UzEpMCcGA1UEAwwgU3VwZXJPZmZpY2UgT25saW5lIFByb2R1Y3Rpb24gQ0ExIjAg
-BgkqhkiG9w0BCQEWE3Nka0BzdXBlcm9mZmljZS5jb20wIBcNMTQwNDA0MDczNDM3
-WhgPMjExNDAzMTEwNzM0MzdaMIGgMQswCQYDVQQGEwJOTzENMAsGA1UECAwET3Ns
-bzENMAsGA1UEBwwET3NsbzEXMBUGA1UECgwOU3VwZXJPZmZpY2UgQVMxNjA0BgNV
-BAMMLVN1cGVyT2ZmaWNlIE9ubGluZSBQcm9kdWN0aW9uIEZlZGVyYXRlZCBMb2dp
-bjEiMCAGCSqGSIb3DQEJARYTc2RrQHN1cGVyb2ZmaWNlLmNvbTCCASIwDQYJKoZI
-hvcNAQEBBQADggEPADCCAQoCggEBAPkQvivcIoLdCKwtlHHhh7aJBuFXjTb8Fxw7
-eqLiIZRvV+XLykGbaHkunlV7oyCxVA07dBY4y3VFO2YYILY8qhjJXvoeOHT9s2l4
-M325b0gJG4G3PSz4Y75DgiVedTI85/iIewCMP+FYe9jiOQCckdbh7jMaLCi2ZTM9
-6XwrnYwZewScLmQ82BNBbS1DcIYQXS6gSN/GqIuMQP9m5OzTfEjpl8L4ngNIqz3S
-LcZ2PWcFGrdg02+68phjEUGeElmf3XkyGcGIjYkgeMZSyoZLPNDsAfNCnuHg3Oeq
-K1Y5jlzT3Ng3yd2DZpiNm3l10MG5/LEfw7k6L8kptrcw+wxMiacCAwEAATANBgkq
-hkiG9w0BAQUFAAOCAQEAofT3yTZytAdRGNOpqNQilQDW34lnEsPzwb2H4UdatNnC
-zoSdl4piiNMXyilJS4gwU4KMfuBDO07i+5Mpt+Mh1SSr7vrDxeU6Sq+vAjuzOE9U
-247PIwLZuxuKD5FLCDomrNyL4+SSzXOyPDkNBezDZmChRTdlylDDRnjxX3SM+tsx
-8Lqw4rwyXCOLMuubWCqYinoFelI3Rj+XDpKxcQ+nwIAhMyoZXL90fEoxdYRoKIy8
-Evvv17LMOFW/CexfThepu3+FbtA3+xPeSidnlXv/9WC3wG1M4nlIiIazbqoXM8gL
-mIUwK8BZ+tIFqCj2oJAJhr68WA4LQ+Kr8k3fnmVGnQ==
------END CERTIFICATE-----
