--- src/DatabaseMirroringProject/App_Data/README.txt
@@ -1,3 +1,0 @@
-Default SuperOfficeFeederatedLogin.crt is for production.
-
-If targeting SOD or Stage, copy and replace the SuperOfficeFeederatedLogin.crt file with the one from the appropriate folder.
