--- src/DatabaseMirroringProject/MirroringClientService.svc
@@ -1,1 +1,0 @@
-<%@ ServiceHost Language="C#" Debug="true" Service="DatabaseMirroringProject.SuperOfficeMirror.MirroringClientService" CodeBehind="MirroringClientService.svc.cs" %>
