--- src/DatabaseMirroringProject/PrivateKey.txt
@@ -1,3 +1,0 @@
-﻿<RSAKeyValue>
-  <YOURPRIVATEKEY \>
-</RSAKeyValue>
