--- src/DatabaseMirroringProject/Web.config
@@ -1,104 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<appSettings>
-		<add key="aspnet:UseTaskFriendlySynchronizationContext" value="true" />		
-		
-		<!-- Obtain an AppId (client_id) at https://docs.superoffice.com/developer-portal/create-app/mirror-app.html -->
-		<add key="SoAppId" value="" />
-		
-		<!--<add key="SoFederationGateway" value="https://sod.superoffice.com/login/" />-->
-		<!--<add key="SoFederationGateway" value="https://qaonline.superoffice.com/login/" />-->
-		<add key="SoFederationGateway" value="https://online.superoffice.com/login/" />
-		
-		<!--SOD-->
-		<!--<add key="SuperIdCertificate" value="16b7fb8c3f9ab06885a800c64e64c97c4ab5e98c" />-->
-		<!--Stage-->
-		<!--<add key="SuperIdCertificate" value="484903a7732b928c7980a39d38e947ea119f1089" />-->
-		<!--Production-->
-		<add key="SuperIdCertificate" value="60b61583dbb2b7eec46fd01ec0056e62451d61f3" />
-		
-		<add key="MirrorDatabaseName" value="YOUR_DATABASE_NAME_[@contextId]" />
-		<add key="ConnectionBase" value="Server=YOUR_SERVER_NAME;User ID=YOUR_DB_USER_NAME;Password=YOUR_DB_USER_PASSWORD" />
-		
-		<!-- Obtain a private RSAXML key at https://docs.superoffice.com/developer-portal/security/new-certificate.html -->
-		<add key="PrivateKeyFile" value="privatekey.txt" />
-		
-		<!--writes events to a log, with a new file each day, located in the system TEMP directory (Mirroring-yyyyMMdd.log)-->
-		<add key="EnableLogging" value="False" />
-	</appSettings>
-	<system.web>
-		<compilation debug="true" targetFramework="4.8" />
-		<httpRuntime targetFramework="4.8" maxRequestLength="2147483647" />
-	</system.web>
-	<system.serviceModel>
-		<behaviors>
-			<serviceBehaviors>
-				<behavior>
-					<!-- To avoid disclosing metadata information, 
-               set the values below to false before deployment -->
-					<serviceMetadata httpGetEnabled="false" httpsGetEnabled="false" />
-					<!-- To receive exception details in faults for debugging purposes, 
-               set the value below to true. Set to false before deployment to 
-               avoid disclosing exception information -->
-					<serviceDebug includeExceptionDetailInFaults="false" />
-				</behavior>
-			</serviceBehaviors>
-		</behaviors>
-		<protocolMapping>
-			<add binding="basicHttpBinding" scheme="http" bindingConfiguration="DbMirroring" />
-		</protocolMapping>
-		<serviceHostingEnvironment 
-			aspNetCompatibilityEnabled="true" 
-			multipleSiteBindingsEnabled="true" />
-		<bindings>
-			<basicHttpBinding>
-				<binding name="DbMirroring" 
-						 maxBufferPoolSize="2147483647" 
-						 maxReceivedMessageSize="2147483647" 
-						 maxBufferSize="2147483647">
-		<!-- Uncomment the following security element if experiencing 'service unavilable', 404 -->
-		<!-- it's unnecessary on many installations, but sometimes required -->
-		<!-- <security mode="Transport">
-				<transport clientCredentialType="None" />
-			 </security> -->
-					<readerQuotas maxDepth="2147483647" 
-								  maxStringContentLength="2147483647" 
-								  maxArrayLength="2147483647" 
-								  maxBytesPerRead="2147483647" 
-								  maxNameTableCharCount="2147483647" />
-				</binding>
-			</basicHttpBinding>
-		</bindings>
-		<services>
-			<service name="DatabaseMirroringProject.SuperOfficeMirror.MirroringClientService">
-				<endpoint binding="basicHttpBinding" 
-						  bindingConfiguration="DbMirroring" 
-						  contract="SuperOffice.Online.Mirroring.Contract.IMirroringClientService" />
-			</service>
-		</services>
-	</system.serviceModel>
-	<system.webServer>
-		<modules runAllManagedModulesForAllRequests="true" />
-		<!--
-        To browse web app root directory during debugging, set the value below to true.
-        Set to false before deployment to avoid disclosing web app folder information.
-      -->
-		<directoryBrowse enabled="true" />
-	</system.webServer>
-	<runtime>
-		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-			<dependentAssembly>
-				<assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-12.0.0.0" newVersion="12.0.0.0" />
-			</dependentAssembly>
-			<dependentAssembly>
-				<assemblyIdentity name="System.Diagnostics.DiagnosticSource" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-4.0.3.1" newVersion="4.0.3.1" />
-			</dependentAssembly>
-			<dependentAssembly>
-				<assemblyIdentity name="System.Runtime.CompilerServices.Unsafe" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-5.0.0.0" newVersion="5.0.0.0" />
-			</dependentAssembly>
-		</assemblyBinding>
-	</runtime>
-</configuration>
