--- src/DatabaseMirroringProject/packages.config
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.ApplicationInsights" version="2.21.0" targetFramework="net48" />
-  <package id="Microsoft.IdentityModel.JsonWebTokens" version="5.4.0" targetFramework="net48" />
-  <package id="Microsoft.IdentityModel.Logging" version="5.4.0" targetFramework="net48" />
-  <package id="Microsoft.IdentityModel.Tokens" version="5.4.0" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net48" />
-  <package id="SuperOffice.Crm.Online.Core" version="4.0.7275.1215" targetFramework="net48" />
-  <package id="SuperOffice.Crm.Online.Mirroring" version="4.0.7275.1215" targetFramework="net48" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
-  <package id="System.Diagnostics.DiagnosticSource" version="5.0.0" targetFramework="net48" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="5.4.0" targetFramework="net48" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="5.0.0" targetFramework="net48" />
-  <package id="System.ServiceModel.Primitives" version="4.5.3" targetFramework="net48" />
-</packages>
