+++ src/DatabaseMirroringWcfService/App_Data/SuperOfficeFederatedLogin.development.crt
