+++ src/DatabaseMirroringWcfService/DatabaseMirroringWcfService.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35527.113 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatabaseMirroringWcfService", "DatabaseMirroringWcfService.csproj", "{0EF793AB-48CB-47A5-AFF7-C82C19DC283E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0EF793AB-48CB-47A5-AFF7-C82C19DC283E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0EF793AB-48CB-47A5-AFF7-C82C19DC283E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0EF793AB-48CB-47A5-AFF7-C82C19DC283E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0EF793AB-48CB-47A5-AFF7-C82C19DC283E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
