+++ src/DatabaseMirroringWcfService/ILLink/ILLink.Descriptors.LibraryBuild.xml
@@ -1,0 +1,8 @@
+﻿<linker>
+  <assembly fullname="System.Diagnostics.DiagnosticSource">
+    <type fullname="System.Diagnostics.Metrics.MetricsEventSource">
+      <!-- Used by System.Private.CoreLib via reflection to init the EventSource -->
+      <method name="GetInstance" />
+    </type>
+  </assembly>
+</linker>
