+++ src/DatabaseMirroringWcfService/PrivateKey.txt
@@ -1,0 +1,3 @@
+﻿<RSAKeyValue>
+  <YOURPRIVATEKEY />
+</RSAKeyValue>
