+++ src/DatabaseMirroringWcfService/Web.Debug.config
