+++ src/DatabaseMirroringWcfService/Web.Release.config
