+++ src/DatabaseMirroringWcfService/Web.config
@@ -3,7 +3,7 @@
   <appSettings>
     <add key="aspnet:UseTaskFriendlySynchronizationContext" value="true" />
     <!-- Obtain an AppId (client_id) at https://docs.superoffice.com/developer-portal/create-app/mirror-app.html -->
-    <add key="SoAppId" value="" />
+    <add key="SoAppId" value="SANITIZED" />
     <!-- Options: Development, Stage, or Production -->
     <add key="Environment" value="Development" />
     <add key="SoFederationGateway" value="https://sod.superoffice.com/login/" />
@@ -18,7 +18,7 @@
     <!--Production-->
     <!--<add key="SuperIdCertificate" value="60b61583dbb2b7eec46fd01ec0056e62451d61f3" />-->
     <add key="MirrorDatabaseName" value="Mirroring_[@contextId]" />
-    <add key="ConnectionBase" value="" />
+    <add key="ConnectionBase" value="Server=YOUR_SERVER_NAME;User ID=YOUR_DB_USER_NAME;Password=YOUR_DB_USER_PASSWORD" />
     <!--writes events to a log, with a new file each day, located in the system TEMP directory (Mirroring-yyyyMMdd.log)-->
     <add key="EnableLogging" value="False" />
   </appSettings>
