+++ src/DatabaseMirroringWcfService/packages.config
@@ -1,0 +1,48 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Azure.Core" version="1.38.0" targetFramework="net48" />
+  <package id="Azure.Identity" version="1.11.4" targetFramework="net48" />
+  <package id="Microsoft.Bcl.AsyncInterfaces" version="8.0.0" targetFramework="net48" />
+  <package id="Microsoft.Bcl.Cryptography" version="8.0.0" targetFramework="net48" />
+  <package id="Microsoft.Bcl.Memory" version="9.0.0" targetFramework="net48" />
+  <package id="Microsoft.Bcl.TimeProvider" version="8.0.1" targetFramework="net48" />
+  <package id="Microsoft.Data.SqlClient" version="6.0.1" targetFramework="net48" />
+  <package id="Microsoft.Data.SqlClient.SNI" version="6.0.2" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Caching.Abstractions" version="8.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Caching.Memory" version="8.0.1" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="8.0.2" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Logging.Abstractions" version="8.0.3" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Options" version="8.0.2" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Primitives" version="8.0.0" targetFramework="net48" />
+  <package id="Microsoft.Identity.Client" version="4.61.3" targetFramework="net48" />
+  <package id="Microsoft.Identity.Client.Extensions.Msal" version="4.61.3" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.Abstractions" version="8.3.1" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.JsonWebTokens" version="8.3.1" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.Logging" version="8.3.1" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.Protocols" version="8.3.1" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.Protocols.OpenIdConnect" version="8.3.1" targetFramework="net48" />
+  <package id="Microsoft.IdentityModel.Tokens" version="8.3.1" targetFramework="net48" />
+  <package id="SuperOffice.Online.Mirroring" version="1.0.0-rc.1" targetFramework="net48" />
+  <package id="SuperOffice.Online.Tokens" version="7.21.0" targetFramework="net48" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
+  <package id="System.ClientModel" version="1.0.0" targetFramework="net48" />
+  <package id="System.Configuration.ConfigurationManager" version="8.0.1" targetFramework="net48" />
+  <package id="System.Diagnostics.DiagnosticSource" version="9.0.1" targetFramework="net48" />
+  <package id="System.Diagnostics.EventLog" version="6.0.0" targetFramework="net48" />
+  <package id="System.IdentityModel.Tokens.Jwt" version="8.3.1" targetFramework="net48" />
+  <package id="System.IO.FileSystem.AccessControl" version="5.0.0" targetFramework="net48" />
+  <package id="System.Memory" version="4.5.5" targetFramework="net48" />
+  <package id="System.Memory.Data" version="1.0.2" targetFramework="net48" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net48" />
+  <package id="System.Security.AccessControl" version="5.0.0" targetFramework="net48" />
+  <package id="System.Security.Cryptography.Pkcs" version="8.0.1" targetFramework="net48" />
+  <package id="System.Security.Cryptography.ProtectedData" version="4.7.0" targetFramework="net48" />
+  <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net48" />
+  <package id="System.ServiceModel.Primitives" version="8.1.1" targetFramework="net48" />
+  <package id="System.Text.Encodings.Web" version="8.0.0" targetFramework="net48" />
+  <package id="System.Text.Json" version="8.0.5" targetFramework="net48" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net48" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
+  <package id="MSBuild.Microsoft.VisualStudio.Web.targets" version="14.0.0.3" targetFramework="net48" />
+</packages>
